# swagger.yaml
openapi: 3.0.0
info:
  title: Attendance Sheet API
  version: 1.0.0
  description: The Attendance Sheet is a structured document designed to systematically record the presence, absence, or tardiness of individuals within a specified group or event. Typically used in educational institutions, workplaces, and various organizational settings, the attendance sheet serves as a vital tool for tracking attendance patterns, facilitating payroll processes, and assessing overall participation
  contact:
    name: ABU TAHER
    url: https://github.com/taher267
    email: abutaher267@gmail.com
servers:
  - description: Dev
    url: "http://localhost:4003/api/v1"
  - description: Live
    url: "https://ecom-server-rw2s.onrender.com/api/v1"
  - description: Ngrok Dynamic
    url: "https://a3e2-103-96-104-72.ngrok-free.app/api/v1"
...
paths:
...
  "/users/{id}":
    get:
      operationId: getUser
      security:
        - bearerAuth: []
      description: Retrieve single user
      tags:
        - "User"
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "200":
          description: A single item details
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Retrieved a data, Successfully!
                  data:
                    type: object
                    allOf:
                      - $ref: "#components/schemas/UserCommon"
                      - $ref: "#components/schemas/Timestamp"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          "$ref": "#/components/responses/500"
    put:
      operationId: updateOrCreateUser
      security:
        - bearerAuth: []
      description: If exist user than update other wise create a user. only update those specified field(name, phone_number, username)
      tags:
        - "User"
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#components/schemas/User"
      responses:
        "200":
          description: Update user info!
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Retrieved a data, Successfully!
                  data:
                    type: object
                    allOf:
                      - $ref: "#components/schemas/UserCommon"
                      - $ref: "#components/schemas/Timestamp"
        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"
    patch:
      operationId: updateUser
      security:
        - bearerAuth: []
      description: only update Status, Roles
      tags:
        - "User"
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        content:
          "application/json":
            schema:
              # oneOf:
              #     - required:
              #         - roles
              #       properties:
              #         roles:
              #           type: array
              #           items:
              #             type: string
              #     - required:
              #         - status
              #       properties:
              #         status:
              #           type: string
              #           enum: ["active", "inactive", "pending"]
              properties:
                roles:
                  type: array
                  items:
                    type: string
                    enum: [super_admin, admin, user, observer]
                status:
                  type: string
                  enum: ["active", "inactive", "pending", "close"]
      responses:
        "200":
          description: Update user info!
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Retrieved a data, Successfully!
                  data:
                    type: object
                    allOf:
                      - $ref: "#components/schemas/UserCommon"
                      - $ref: "#components/schemas/Timestamp"
        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"

    delete:
      operationId: deleteUser
      security:
        - bearerAuth: []
      description: Delete single data
      tags:
        - "User"
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "200":
          description: Delete single item details
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 202
                  message:
                    type: string
                    example: Data has been deleted, Successfully!
        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"
...
  "/departments/{id}":
    get:
      operationId: Department
      description: "Retrive Department"
      tags:
        - "Department"
      parameters:
        - $ref: "#/components/parameters/idPathParam"

      responses:
        "200":
          description: Fetching a Department
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Added department, Successfully
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: Hadia Shop
                      establishment_id:
                        type: string
                        example: 6568c4210a26245b8cb27313
        "401":
          "$ref": "#/components/responses/401"
        "403":
          "$ref": "#/components/responses/403"
        "404":
          "$ref": "#/components/responses/404"
        "500":
          "$ref": "#/components/responses/500"
    patch:
      operationId: updateEstablishment
      security:
        - bearerAuth: []
      description: "Update Establishment"
      tags:
        - "Establishment"
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                name:
                  description: Establishment name must be unique!
                  type: string
                  example: Hadia Shop

      responses:
        "200":
          description: Update Department
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Update department, Successfully
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: Hadia Shop
                      establishment_id:
                        type: string
                        example: 6568c4210a26245b8cb27313
    delete:
      operationId: DeleteDepartment
      security:
        - bearerAuth: []
      description: "Delete Department"
      tags:
        - "Department"
      parameters:
        - $ref: "#/components/parameters/idPathParam"

      responses:
        "202":
          description: Department Deleting
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 202
                  message:
                    type: string
                    example: Department has been deleted!

        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    idPathParam:
      in: path
      name: id
      required: true
      description: Data ID of single item Retrieve
      schema:
        type: string
        example: "64d268cdeef8c84bb18912db"
...
  responses:
    "500":
      description: "Internal Server Error"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later
    "400":
      description: "Bad Request"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
                    in:
                      type: string
                      enum: [query, body]
                      example: body
    "401":
      description: Unauthorized
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: You don't have the right permissions
    "403":
      description: Unauthenticated
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthenticated
              message:
                type: string
                example: You don't have the right permissions
    "404":
      description: Resource Not Found
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Requested resource not found